requirements.txt
# ==================== PRODUCCIÓN ====================
# Core API
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AWS
boto3==1.34.0

# Utilities
python-multipart==0.0.6
PyJWT==2.8.0
email-validator==2.1.0
mangum==0.17.0

# ==================== DESARROLLO ====================
# Testing (opcional - solo para desarrollo)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2